move probe location build after instrumentation
move summaryBuilder
move and rename executeScript
