+++ dd-java-agent/agent-debugger/src/main/java/com/datadog/debugger/probe/MetricProbe.java
@@ -6,6 +6,7 @@ import com.datadog.debugger.instrumentation.MetricInstrumentor;
 import datadog.trace.bootstrap.debugger.DiagnosticMessage;
 import datadog.trace.bootstrap.debugger.MethodLocation;
 import datadog.trace.bootstrap.debugger.ProbeId;
+import datadog.trace.bootstrap.debugger.SummaryBuilder;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
@@ -67,6 +68,11 @@ public class MetricProbe extends ProbeDefinition {
       List<String> probeIds) {
     new MetricInstrumentor(this, classLoader, classNode, methodNode, diagnostics, probeIds)
         .instrument();
+  }
+
+  @Override
+  public SummaryBuilder getSummaryBuilder() {
+    return null;
   }
 
   public static Builder builder() {
